レイアウト部分を共通コンポーネントへ抜き出し
